Fix "Lobby initialization" unit test
The Docker image we were using to build and run tests was based on Alpine Linux, which does not ship the necessary packages to allow `java.awt.Font` to load a font correctly, which caused the test to fail.

This change should not increase the overall image size because we are using a multi-stage Docker build.