Fix GuiModule losing inventory references after opening them a second time
This bug never surfaced before because we used `InventoryCondition`s, which are bound to the `Inventory`. When they were removed from Minestom, we switched to processing `InventoryPreClickEvent`s manually, which revealed a bug: when a player reopened an inventory that had already been created by its containing `Menu`, we used a cached version, but didn't add it back to the `inventories` map, meaning the GuiModule didn't propagate any clicks made in that inventory to their handlers.